Added new classes to facilitate model parameter fitting.
