+++ python/ctisim/__init__.py
@@ -1,7 +1,7 @@
 from ctisim.core import SegmentSimulator
 from ctisim.core import SerialTrap
 from ctisim.core import SerialRegister
-from ctisim.core import ReadoutAmplifier
+from ctisim.core import OutputAmplifier
 from ctisim.core import ImageSimulator
 from ctisim.utils import ITL_AMP_GEOM
 from ctisim.utils import E2V_AMP_GEOM
