+++ python/ctisim/mcmc.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from ctisim import SerialTrap, SerialRegister, ReadoutAmplifier, SegmentSimulator
+from ctisim import SerialTrap, SerialRegister, OutputAmplifier, SegmentSimulator
 
 class TrapModelFitting:
     """Object to control MCMC parameter fitting of a serial trapping model."""
@@ -27,7 +27,8 @@ class TrapModelFitting:
         else:
             return -np.inf
             
-    def loglikelihood(self, trap_params, flux_array, data_rows, error, trap_locations, biasdrift_params=None):
+    def loglikelihood(self, trap_params, flux_array, data_rows, error, 
+                      trap_locations, biasdrift_params=None):
         """Calculate log likelihood for model with given parameters."""
         
         ctiexp, densityfactor, tau, trapsize = trap_params
@@ -36,8 +37,8 @@ class TrapModelFitting:
         
         start = self.amp_geom['num_serial_prescan'] + self.amp_geom['ncols']
         stop = start+self.num_oscan_pixels
-        model_rows = self.simplified_model(model_params, flux_array, trap_locations, self.amp_geom,
-                                          biasdrift_params=biasdrift_params)
+        model_rows = self.simplified_model(model_params, flux_array, trap_locations, 
+                                           self.amp_geom, biasdrift_params=biasdrift_params)
         
         model = model_rows[:, start:stop]
         data = data_rows[:, start:stop]
@@ -45,7 +46,8 @@ class TrapModelFitting:
         
         return -0.5*(np.sum(inv_sigma2*(data-model)**2.))
         
-    def logprobability(self, trap_params, flux_array, meanrows, error, trap_locations, biasdrift_params=None):
+    def logprobability(self, trap_params, flux_array, meanrows, error, 
+                       trap_locations, biasdrift_params=None):
         """Calculate sum of log likelihood and log prior."""
         
         lp = self.logprior(trap_params)
@@ -57,7 +59,8 @@ class TrapModelFitting:
             return result
         
     @staticmethod
-    def simplified_model(trap_params, flux_array, trap_location, amp_geom, biasdrift_params=None):
+    def simplified_model(trap_params, flux_array, trap_location, amp_geom, 
+                         biasdrift_params=None):
         """Generate simulated data for given serial trap parameters."""
     
         cti, df, tau, trap_size = trap_params
@@ -69,10 +72,11 @@ class TrapModelFitting:
 
         trap = SerialTrap(df, tau, trap_size, location=trap_location)
         serial_register = SerialRegister(ncols+num_serial_prescan, cti, trap=trap)
-        readout_amplifier = ReadoutAmplifier(0.0, gain=1.0, offset=0.0, biasdrift_params=biasdrift_params)
+        output_amplifier = OutputAmplifier(0.0, gain=1.0, offset=0.0, 
+                                             biasdrift_params=biasdrift_params)
         flat = SegmentSimulator((len(flux_array), ncols), num_serial_prescan)
         flat.ramp_exp(flux_array)
 
-        return readout_amplifier.serial_readout(flat, serial_register, 
+        return output_amplifier.serial_readout(flat, serial_register, 
                                                 num_serial_overscan=num_serial_overscan, 
                                                 num_parallel_overscan=num_parallel_overscan)
